+++ src/data-layer/adapters/Payload.ts
@@ -1,0 +1,13 @@
+import { testPayloadObject } from '@/test-config/utils'
+import configPromise from '@payload-config'
+import { Payload, getPayload } from 'payload'
+
+let payloadConfig = await getPayload({
+  config: configPromise,
+})
+
+if (process.env.NODE_ENV === 'test') {
+  payloadConfig = testPayloadObject
+}
+
+export const payload: Payload = payloadConfig
