+++ src/data-layer/services/AuthService.test.ts
@@ -1,7 +1,7 @@
+import { authenticationCreateMock } from '@/test-config/mocks/Authentication.mock'
+import { userMock } from '@/test-config/mocks/User.mock'
+import { clearCollection, testPayloadObject } from '@/test-config/utils'
 import dotenv from 'dotenv'
-import { authenticationCreateMock } from 'tests/mocks/Authentication.mock'
-import { userMock } from 'tests/mocks/User.mock'
-import { clearCollection, testPayloadObject } from 'tests/utils'
 import AuthService from './AuthService'
 
 dotenv.config()
