+++ src/data-layer/services/UserService.test.ts
@@ -1,6 +1,6 @@
+import { userMock } from '@/test-config/mocks/User.mock'
+import { clearCollection, testPayloadObject } from '@/test-config/utils'
 import dotenv from 'dotenv'
-import { userMock } from 'tests/mocks/User.mock'
-import { clearCollection, testPayloadObject } from 'tests/utils'
 import UserService from './UserService'
 
 dotenv.config()
